Where can I find the heaven? I don't know

+++ src/main/java/com/kozi/koziAPI/repository/CategoriasRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.kozi.koziAPI.model.Categorias;
 
 @Repository
-public interface CategoriasRepository extends JpaRepository<Categorias, Integer>{
+public interface CategoriasRepository extends JpaRepository<Categorias, Long>{
 
 }
