+++ src/main/java/com/kozi/koziAPI/repository/ComunaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.kozi.koziAPI.model.Comuna;
 
 @Repository
-public interface ComunaRepository extends JpaRepository<Comuna, Integer> {
+public interface ComunaRepository extends JpaRepository<Comuna, Long> {
 
 }
