+++ src/main/java/com/kozi/koziAPI/repository/EnvioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.kozi.koziAPI.model.Envio;
 
 @Repository
-public interface EnvioRepository extends JpaRepository<Envio, Integer>{
+public interface EnvioRepository extends JpaRepository<Envio, Long>{
 
 }
