+++ src/main/java/com/kozi/koziAPI/repository/PagoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kozi.koziAPI.model.Pago;
 
-public interface PagoRepository extends JpaRepository<Pago, Integer>{
+public interface PagoRepository extends JpaRepository<Pago, Long>{
     
 }
